refactor: createPlaylistの通信部分モジュール化
